requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
langgraph==0.0.26
langchain==0.0.350
langchain-openai==0.0.2
python-dateutil==2.8.2
requests==2.31.0